com/fuyin/ywbd/base/requirements.txt
openai==1.12.0
python-dotenv
pdfminer.six
nltk==3.8.1
chromadb==0.4.15
gradio